requirements.txt
Flask
mysql-connector-python
nltk
fuzzywuzzy
python-Levenshtein
